+++ packages/dialect-react-ui/CHANGELOG.md
@@ -1,7 +1,24 @@
 # CHANGELOG
 
 ## [UNRELEASED]
+
+- added logic to notfication ui to display new notification badge
+- fix: open thread if user entered address with whom thread already exists
+- fix: do not fetch handles and avatars twice
+
+
+## [0.8.9] - 2022-06-07
+
+- fix: add missing cardinal dependency
+
+## [0.8.8] - 2022-06-01
+
+- Fix published version build
+
+## [0.8.7] - 2022-06-01
+
 - Add solana name service support
+- Add polling props to notfication button
 
 ## [0.8.6] - 2022-05-11